backend/requirements.txt
Flask
python-dotenv
requests
Flask-CORS
google-generativeai
chromadb